refactor test_utils to have submodules
